Remove luna_minclient_dir
This patch removes the need to specify the top level directory in the
tarball.  Instead it just strips the top level without having to know
the name of it.

This simplifies the deployment some because a user does not need to
inspect the tarball contents to get the name of this directory.